refactor(appeals/api): appeals api directory restructuring following the big split
