+++ appeals/api/index.d.ts
@@ -1,4 +1,4 @@
-import * as Appeals from './src/server/appeals/appeals';
+import * as Appeals from './src/server/endpoints/appeals';
 import * as Schema from './src/database/schema';
 
 export { Appeals, Schema };
