+++ appeals/api/src/server/build-app.js
@@ -3,7 +3,7 @@ import compression from 'compression';
 import express from 'express';
 import helmet from 'helmet';
 import morgan from 'morgan';
-import { appealsRoutes } from './appeals/appeals.routes.js';
+import { appealsRoutes } from './endpoints/appeals.routes.js';
 import { defaultErrorHandler, stateMachineErrorHandler } from './middleware/error-handler.js';
 import versionRoutes from './middleware/version-routes.js';
 import BackOfficeAppError from './utils/app-error.js';
