+++ appeals/api/src/server/endpoints/appeals/appeals.routes.js
@@ -1,6 +1,6 @@
 import { Router as createRouter } from 'express';
 import { asyncHandler } from '../../middleware/async-handler.js';
-import { documentsRoutes } from '../../documents/documents.routes.js';
+import { documentsRoutes } from '../documents/documents.routes.js';
 import {
 	getAppealById,
 	getAppeals,
