+++ appeals/api/src/server/endpoints/appellant-case-incomplete-reasons/__tests__/appellant-case-incomplete-reasons.test.js
@@ -1,6 +1,6 @@
 import supertest from 'supertest';
 import { app } from '../../../app-test.js';
-import { appellantCaseIncompleteReasons } from '../../tests/data.js';
+import { appellantCaseIncompleteReasons } from '../../../tests/data.js';
 import { ERROR_FAILED_TO_GET_DATA, ERROR_NOT_FOUND } from '../../constants.js';
 
 const { databaseConnector } = await import('../../../utils/database-connector.js');
